data-lake-pipeline/requirements.txt
apache-airflow==2.1.2
pyspark==3.2.0
boto3==1.18.2
kafka-python==2.0.2
awscli==1.22.5
pandas==1.3.3
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pyyaml==5.4.1
minio==7.1.8